+++ pebblemon/src/c/modules/pokemon/menu.h
@@ -66,4 +66,4 @@ void draw_player_battle_frame(GBC_Graphics *graphics);
 void draw_battle_frames(GBC_Graphics *graphics);
 void draw_enemy_hp_bar(GBC_Graphics *graphics, uint16_t max_health, uint16_t cur_health);
 void draw_player_hp_bar(GBC_Graphics *graphics, uint16_t max_health, uint16_t cur_health);
-void draw_exp_bar(GBC_Graphics *graphics, uint16_t max_exp, uint16_t cur_exp);
+void draw_exp_bar(GBC_Graphics *graphics, int max_exp, int cur_exp);
